user_service/requirements.txt
flask==2.3.2
python-dotenv